Se agregar correcciones a SimplexBuilder y a DispositivosHelper

+++ SGE.Core/Entidades/Dispositivo.cs
@@ -8,7 +8,7 @@ namespace SGE.Core.Entidades
         public string Tipo { get; set; }
         public string EsInteligente { get; set; }
         public string EsBajoConsumo { get; set; }
-        public decimal Consumo { get; set; }
+        public double Consumo { get; set; }
         public List<string> Caracteristicas { get; set; }
     }
 }
